requirements.txt
fastapi
selenium
webdriver-manager
pydantic
pandas
uvicorn
pillow
